Assignment 12/requirements.txt
Django>=3.2
 djangorestframework
django-filter
